feat: 🎸 Seperate auth check
